requirements.txt
django
numpy
pandas
scipy
numpy-stl
pillow
opencv-contrib-python
django-cleanup